requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
ecdsa==0.18.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
uvicorn==0.27.1
